Moved FeruchemyChargeThread.java to /client/utils from /common/utils
