+++ src/feruchemy/java/leaf/cosmere/feruchemy/client/FeruchemyClientEvents.java
@@ -4,23 +4,17 @@
 
 package leaf.cosmere.feruchemy.client;
 
-import leaf.cosmere.api.CosmereAPI;
-import leaf.cosmere.api.Metals;
 import leaf.cosmere.api.helpers.EntityHelper;
-import leaf.cosmere.common.items.ChargeableMetalCurioItem;
 import leaf.cosmere.common.registry.AttributesRegistry;
 import leaf.cosmere.feruchemy.common.Feruchemy;
 import leaf.cosmere.feruchemy.common.manifestation.FeruchemyAtium;
-import leaf.cosmere.feruchemy.common.utils.FeruchemyChargeThread;
-import net.minecraft.client.Minecraft;
+import leaf.cosmere.feruchemy.client.utils.FeruchemyChargeThread;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.RenderLivingEvent;
 import net.minecraftforge.client.event.RenderNameTagEvent;
 import net.minecraftforge.client.event.RenderPlayerEvent;
-import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.eventbus.api.Event;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
